Loans replaced with loan collections.
This means that user can build any combination of loans into a combined loan-collection. This is useful to test the curves of real world situations where people often have multiple mortgages on a single property. The current Icelandic landscape also includes mixed loans of indexed/unindexed parts and this would be a requirement to calculating them.

This now adds the need to show collection breakdowns as small multiples.

As it turns out I am starting to regret some design decisions, it would have been much less complicated to classify the loan collection and loans into a model. The current structure made more sense when the form was static. I've added a bunch of helper methods to reduce the work of dealing with it, but I may need to refactor the whole thing if any more complexity is added.